test(build): move the template contract tests to their own file
These assert invariants of the cookiecutter template itself — that every
context key a template reads is declared, and that pubspec.yaml parses
unrendered and renders to valid YAML — none of which is Linux-specific.
They only lived in the Linux file because that is where the bugs that
motivated them were found.

Each file keeps its own copy of the small render helper rather than
introducing a conftest, matching how the other suites here are written.
