+++ sdk/python/packages/flet-cli/tests/test_build_linux.py
@@ -15,9 +15,7 @@ Two halves, both driven against the real in-repo build template:
 """
 
 import argparse
-import json
 import os
-import re
 from pathlib import Path
 from types import SimpleNamespace
 from typing import Any, Optional
@@ -461,70 +459,3 @@ class TestDesktopEntryTemplate:
         )
         assert 'Exec="weird\\\\\\"name" %U' in content
         assert "Categories=Game;Fun;" in content
-
-
-class TestBuildTemplateContract:
-    """
-    Invariants of the build template itself, independent of any one platform.
-    """
-
-    @staticmethod
-    def _declared_keys() -> set:
-        """The context keys `cookiecutter.json` declares."""
-        with open(BUILD_TEMPLATE_DIR.parent / "cookiecutter.json") as f:
-            return set(json.load(f))
-
-    @staticmethod
-    def _referenced_keys() -> dict:
-        """Every `cookiecutter.<key>` referenced by a template, by file."""
-        found = {}
-        for path in BUILD_TEMPLATE_DIR.rglob("*"):
-            if not path.is_file():
-                continue
-            try:
-                text = path.read_text()
-            except UnicodeDecodeError:
-                continue  # binary asset, e.g. images/icon.png
-            # The negative lookahead skips method calls such as
-            # `cookiecutter.items()` in .vars, which are not context keys.
-            for key in re.findall(
-                r"cookiecutter\.([a-zA-Z_][a-zA-Z0-9_]*)\b(?!\s*\()", text
-            ):
-                found.setdefault(key, set()).add(path.name)
-        return found
-
-    def test_every_referenced_key_is_declared(self):
-        """
-        A key a template reads but `cookiecutter.json` omits is dropped from
-        the context, so the template silently renders it empty.
-        """
-        declared = self._declared_keys()
-        undeclared = {
-            key: sorted(files)
-            for key, files in self._referenced_keys().items()
-            if key not in declared
-        }
-        assert not undeclared, f"referenced but not declared: {undeclared}"
-
-    def test_pubspec_parses_while_unrendered(self):
-        """
-        `.github/scripts/patch_pubspec_version.py` loads this file as YAML
-        before cookiecutter ever runs, so an unquoted `{{` — which YAML reads
-        as a flow mapping — fails the release build.
-        """
-        with open(TEMPLATE_PUBSPEC) as f:
-            assert yaml.safe_load(f), "template pubspec.yaml must parse as YAML"
-
-    @pytest.mark.parametrize(
-        "description",
-        ["plain", "Bob's tool", 'has "quotes"', "back\\slash", "multi\nline\ttab"],
-    )
-    def test_rendered_pubspec_is_valid_yaml(self, description):
-        """A description of any shape must still render a parseable pubspec."""
-        rendered = _render_template(
-            TEMPLATE_PUBSPEC,
-            project_name="test_app",
-            pubspec_description=BaseBuildCommand.escape_single_quoted_yaml(description),
-        )
-        parsed = yaml.safe_load(rendered)
-        assert parsed["description"] == re.sub(r"[\x00-\x1f]", " ", description)
